Refactor to allow dependency injection
